pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clip-quality-env"
version = "1.0.0"
description = "ClipQualityEnv — OpenEnv RL environment for clip-quality analysis and classification"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "requests>=2.25.0",
    "websockets>=12.0",
    "gradio>=4.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["clip_quality_env", "server"]
